refactor(ci): move the changed-files predicates to their own module
The seventeen shouldXxx predicates were module-private to the pull-request workflow. That
blocked extracting the remaining job groups: a group asking whether the backend changed would
have had to import from the workflow that imports it back.

They move to workflows/groups/changed-files.ts, which becomes the single place where 'this
change can only affect X' is expressed. Narrowing or widening one of them is a deliberate
change to CI coverage, and it now shows up as a diff in a file that has no other subject.

workflow-pull-requests.ts drops from 961 to 553 lines across the three extractions so far.
Generated configuration is byte-for-byte identical: all 50 fixtures are untouched.
